[Next Gen] refactor(atelier): split SSR statement IR nodes into render_ir_ssr
